backend/requirements.txt
Flask
flask-cors
flask-bcrypt
flask-sqlalchemy
googletrans==3.1.0a0
pillow
wordcloud
pytesseract
gunicorn
requests
better-profanity
google-generativeai
google-cloud-aiplatform
google-cloud-speech
google-cloud-texttospeech
openai-whisper
python-dotenv
gtts
firebase-admin